refactor: changed interaction with env files
