+++ .env.example
@@ -1,0 +1,7 @@
+VITE_PROJECT_KEY=name-project
+VITE_CLIENT_ID=string
+VITE_CLIENT_SECRET=string
+VITE_SCOPES=manage_project:name-project
+VITE_API_URL=https://api.europe-west1.gcp.commercetools.com
+VITE_AUTH_URL=https://auth.europe-west1.gcp.commercetools.com
+VITE_NUM=number
