+++ .gitignore
@@ -1,3 +1,4 @@
+.env
 .env.local
 # Logs
 logs
