+++ src/server/api.ts
@@ -1,6 +1,6 @@
+import { createRequestBuilder } from '@/server/client'
 import type { ClientResponse } from '@commercetools/platform-sdk'
 import type { ProductPagedQueryResponse } from '@commercetools/platform-sdk/dist/declarations/src/generated/models/product'
-import { createRequestBuilder } from '@/server/client.ts'
 
 /**
  *
